Move core logic to AbilityModel
